requirements.txt
selenium
flask
pandas
matplotlib
python-dotenv